+++ src/onegov.core/onegov/core/security/rules.py
@@ -1,4 +1,7 @@
+from onegov.core.cronjobs import Job
 from onegov.core.framework import Framework
+from onegov.core.security import Public
+from webob.exc import HTTPException
 
 
 @Framework.permission_rule(model=object, permission=object, identity=None)
@@ -36,3 +39,30 @@ def has_permission_logged_in(app, identity, model, permission):
             return False
 
     return permission in getattr(app.settings.roles, identity.role)
+
+
+@Framework.permission_rule(
+    model=HTTPException,
+    permission=Public,
+    identity=None)
+def may_view_http_errors_not_logged_in(app, identity, model, permission):
+    """ HTTP errors may be viewed by anyone, regardeless of settings.
+
+    This is important, otherwise the HTTPForbidden/HTTPNotFound views
+    will lead to an exception if the user does not have the ``Public``
+    permission.
+
+    """
+    return True
+
+
+@Framework.permission_rule(
+    model=Job,
+    permission=Public,
+    identity=None)
+def may_view_cronjobs_not_logged_in(app, identity, model, permission):
+    """ Cronjobs are run anonymously from a thread and need to be excluded
+    from the permission rules as a result.
+
+    """
+    return True